requirements.txt
streamlit
google-generativeai
faiss-cpu
pypdf
numpy